refactor: 사용하는 모델 정리
