+++ app/src/main/java/com/boosters/promise/data/friend/source/local/FriendLocalDataSource.kt
@@ -1,0 +1,9 @@
+package com.boosters.promise.data.friend.source.local
+
+interface FriendLocalDataSource {
+
+    suspend fun getFriends(): List<FriendEntity>
+
+    suspend fun insertFriend(user: FriendEntity)
+
+}
