--- app/src/main/java/com/boosters/promise/data/invite/source/local/FriendLocalDataSource.kt
@@ -1,9 +1,0 @@
-package com.boosters.promise.data.invite.source.local
-
-interface FriendLocalDataSource {
-
-    suspend fun getFriends(): List<UserEntity>
-
-    suspend fun insertFriend(user: UserEntity)
-
-}
