--- app/src/main/java/com/boosters/promise/data/invite/source/local/FriendLocalDataSourceImpl.kt
@@ -1,17 +1,0 @@
-package com.boosters.promise.data.invite.source.local
-
-import javax.inject.Inject
-
-class FriendLocalDataSourceImpl @Inject constructor(
-    private val userDao: UserDao
-) : FriendLocalDataSource {
-
-    override suspend fun getFriends(): List<UserEntity> {
-        return userDao.getFriends()
-    }
-
-    override suspend fun insertFriend(user: UserEntity) {
-        userDao.insertFriend(user)
-    }
-
-}
